Changed entering/exiting desert
Fixes #10 - You will now enter the desert at a random spot.
To exit the desert you need to find a magical portal that has a chance
to either send you to Hogwarts or to another random spot at the desert.

The magical portals move around and sort of blend with sandstorm, they
can be noticed by the blue magical sparkles.